Added CpuTempToCom7 installer.
When I first did this I was doing in in an uno.
The system assigns COM 1 to an internal USB device i have, COM 3 to the uno, COM 6 to the ESP32 and COM7 to the nano.
Since I will be using a nano for this, I made a new program for it to work on COM7. Since I did it for myself, i figured I might aswell do an installer and upload it here.